lib: Turn GeocodeLocation into a proper GObject
While I made this change under the wrong impression that gjs is not very
happy about GeocodeLocation not being a gobject, this IMO is still a
good change and I put some time into it so I'm sending this patch.

I also took the liberty of changing all tabs to spaces in
geocode-location.[ch]. Looking at geocode-ipclient.c, this seem to be
the desired coding-style.

https://bugzilla.gnome.org/show_bug.cgi?id=696528