lib: Use libsoup directly for HTTP GET requests
We should be doing this anyways, libsoup being our direct dependency but
this patch also avoids the issue of reverse geocoding testcase hanging
becuase of some gvfs bug.

At the same time, this change reveals that our reverse geocoding API is
not working anymore since the Yahoo web API we rely on does not seem to
be available anymore.

https://bugzilla.gnome.org/show_bug.cgi?id=702775