Make include directory API versioned
Note that this will break users of geocode-glib which did not
use the geocode-glib/ subdirectory prefix in the include path
(like geoclue2), according to how the pkg-config file was layed
out before.

Remaining cleanup is to unify the internal include paths inside
geocode-glib to be either "..." or <geocode-glib/...>.

While forcing geocode-glib uses to update their include paths
and internally cleaning up the include paths, this might also
be a good time to drop the filename prefix on header files
since both subdirectory and filename prefix in include paths
seems redundant.

https://bugzilla.gnome.org/show_bug.cgi?id=703571