lib: Better name/description for street numbers
Currently we'd just show the house number in results for street number
searches, e.g Search for "Otavantie 2" would result in places named "2".
This patch fixes this issue by combining house number and street name.

Since format is not the same in all locales, we allow the translators to
format the string.

https://bugzilla.gnome.org/show_bug.cgi?id=705518