forward: switch to jsonv2 to include place_type
The json format does not include the place_rank attribute.
The place_rank attribute is needed to establish what kind of
boundary we are dealing with.

Switching to jsonv2 gives us the place_rank at the cost of having
to rename the class attribute as category.

This patch also fix a typo where boundary was spelled
as boundry. Ensuring that the missing place_rank attribute
never was detected. If fixed in a stand alone patch it exposes
us to a commit that introduces a segfault.

https://bugzilla.gnome.org/show_bug.cgi?id=723463