forward, reverse: Use https and fix the cache
At the moment the cache for queries does not work. This is because
we are using a query with 'http://' to look for cached contents.
But when we are saving we are using 'https://' since that is what
the nominatim server redirects us to.

Regardless of that, moving to 'https://' seems like a good thing.