tests: Port test-gcglib to optionally use GeocodeNominatim offline
GeocodeNominatim allows us to override its virtual methods for network
queries, so do that in a new GeocodeNominatimTest subclass, which loads
its results from a local set of results added by the test harness.

Port the existing test-gcglib tests to optionally use this new backend,
rather than the default GNOME Nominatim backend, to run the tests
offline. This means that they can reliably be run on CI or build
infrastructure with no internet access; and can be run with internet
access on developer machines (for example).

Based on a patch by Aleksander Morgado
<aleksander.morgado@collabora.co.uk>.

https://bugzilla.gnome.org/show_bug.cgi?id=756311