geocode-backend: Add support for multiple reverse results
Some geocoding APIs (such as the Google Geocoding API) can return
multiple results for a reverse resolution request, such as the building
at that location, the street it is on, the town it is in, etc.

While we currently only support Nominatim, other backends may be added
(potentially out of tree) in future. Change the GeocodeBackend API to
allow them to return multiple reverse results.

This does not change the GeocodeReverse API, which returns the first
result in the list, which should be the most relevant one.

https://developers.google.com/maps/documentation/geocoding/intro#ReverseGeocoding

https://bugzilla.gnome.org/show_bug.cgi?id=756311