geocode-glib: Accept a hash table for GeocodeBackend.reverse_resolve()
Instead of only accepting a GeocodeLocation, accept a hash table of
string keys and GValues. Typically, this will contain `lat` and `lon`
entries which represent the GeocodeLocation. But in future, it could
contain other keys to look up a GeocodePlace by its backend-specific ID,
for example; or to limit the search results by place type.

Whether the keys are static or non-static is decided by the caller.

This changes the API in GeocodeBackend (which is not stable yet), but
not in GeocodeReverse, so it’s not an API break.

https://bugzilla.gnome.org/show_bug.cgi?id=756311