reverse: Fix type of lat/lon parameters to backend reverse query method
The backend methods are documented as taking parameters which follow the
Telepathy Location mapping
(https://telepathy.freedesktop.org/spec/Connection_Interface_Location.html#Mapping:Location),
or alternatively XEP-0080.

Both of these specify that the ‘lat’ and ‘lon’ parameters are numeric,
not numeric strings. Since the GeocodeBackend interface has not been
released yet, fix the implementation to use a GValue holding a gdouble,
rather than a GValue holding a string representation of a double.

https://bugzilla.gnome.org/show_bug.cgi?id=774631