ci: Fix undoing delangification when package isn't available
We need to install or reinstall glib2 and glibc packages after removing
/etc/rpm/macros.image-language-conf or we won't have the translation
files we want installed. But in some cases, the package might not be
available for reinstall, so try to install the dependencies and only
reinstall the package if a newer version didn't get installed.

Installed package XXX (from koji-override-1) not available