nominatim: Validate response as UTF-8
Previously, the bytes read from the server's response were assumed to be
valid UTF-8.  Particularly since we are copying it anyway, validate it
at the same time.

(In practice this is probably a non-issue because it just ends up
getting passed to json_parser_load_from_data() which validates its input
as UTF-8 anyway.)